+++ tsdef/RemoteTrackPublication.d.ts
@@ -5,6 +5,9 @@ import { TwilioError } from './TwilioError';
 
 export class RemoteTrackPublication extends TrackPublication {
   isSubscribed: boolean;
+  /**
+   * @deprecated Use (.track.switchOffReason !== Track.SwitchOffReason.DISABLED_BY_PUBLISHER)
+   */
   isTrackEnabled: boolean;
   kind: Track.Kind;
   publishPriority: Track.Priority;
@@ -13,7 +16,7 @@ export class RemoteTrackPublication extends TrackPublication {
   on(event: 'publishPriorityChanged', listener: (priority: Track.Priority) => void): this;
   on(event: 'subscribed', listener: (track: RemoteTrack) => void): this;
   on(event: 'subscriptionFailed', listener: (error: TwilioError) => void): this;
-  on(event: 'trackSwitchedOff', listener: (track: RemoteTrack) => void): this;
+  on(event: 'trackSwitchedOff', listener: (track: RemoteTrack, switchOffReason: Track.SwitchOffReason) => void): this;
   on(event: 'trackSwitchedOn', listener: (track: RemoteTrack) => void): this;
   on(event: 'unsubscribed', listener: (track: RemoteTrack) => void): this;
   on(event: string, listener: (...args: any[]) => void): this;