+++ tsdef/RemoteVideoTrackPublication.d.ts
@@ -8,6 +8,14 @@ export class RemoteVideoTrackPublication extends RemoteTrackPublication {
 
   on(event: 'subscribed', listener: (track: RemoteVideoTrack) => void): this;
   on(event: 'subscriptionFailed', listener: (error: TwilioError) => void): this;
+  /**
+   * @deprecated Use "trackSwitchedOff" (.switchOffReason === Track.SwitchOffReason.DISABLED_BY_PUBLISHER) instead
+   */
+  on(event: 'trackDisabled', listener: () => void): this;
+  /**
+   * @deprecated Use "trackSwitchedOn" instead
+   */
+  on(event: 'trackEnabled', listener: () => void): this;
   on(event: 'trackSwitchedOff', listener: (track: RemoteVideoTrack) => void): this;
   on(event: 'trackSwitchedOn', listener: (track: RemoteVideoTrack) => void): this;
   on(event: 'unsubscribed', listener: (track: RemoteVideoTrack) => void): this;